[wrangler] Move kv/ commands to src/commands/kv/ (Phase 2a)
Move KV namespace commands into the src/commands/ directory structure.

Files moved:
- src/kv/index.ts → src/commands/kv/index.ts
- src/kv/helpers.ts → src/commands/kv/helpers.ts

Imports updated in:
- src/index.ts
- src/sites.ts
- src/deployment-bundle/bindings.ts
- src/commands/delete.ts
- src/__tests__/kv/*.test.ts
- src/__tests__/delete.test.ts
- src/__tests__/helpers/mock-kv.ts

No logic changes - only file moves and import updates.
All 2827 tests pass.
