+++ packages/wrangler/src/__tests__/kv/key.test.ts
@@ -11,7 +11,7 @@ import { createFetchResult, msw } from "../helpers/msw";
 import { runInTempDir } from "../helpers/run-in-tmp";
 import { runWrangler } from "../helpers/run-wrangler";
 import { wranglerKVConfig } from "./constant";
-import type { KeyValue, NamespaceKeyInfo } from "../../kv/helpers";
+import type { KeyValue, NamespaceKeyInfo } from "../../commands/kv/helpers";
 
 describe("kv", () => {
 	mockAccountId();
