+++ packages/wrangler/src/deployment-bundle/bindings.ts
@@ -8,12 +8,12 @@ import {
 	UserError,
 } from "@cloudflare/workers-utils";
 import { fetchResult } from "../cfetch";
+import { createKVNamespace, listKVNamespaces } from "../commands/kv/helpers";
 import { createD1Database } from "../d1/create";
 import { listDatabases } from "../d1/list";
 import { getDatabaseInfoFromIdOrName } from "../d1/utils";
 import { prompt, select } from "../dialogs";
 import { isNonInteractiveOrCI } from "../is-interactive";
-import { createKVNamespace, listKVNamespaces } from "../kv/helpers";
 import { logger } from "../logger";
 import * as metrics from "../metrics";
 import {
